tm: Move page data structure into separate source file
