+++ modules/tm/src/vmem_tx.c
@@ -18,222 +18,11 @@
  */
 
 #include "vmem_tx.h"
-#include <errno.h>
 #include <picotm/picotm-error.h>
-#include <picotm/picotm-lib-rwstate.h>
 #include <stdlib.h>
 #include <string.h>
-#include "block.h"
-#include "frame.h"
 #include "picotm/picotm-tm.h"
-#include "vmem.h"
-
-/*
- * |struct tm_page|
- */
-
-enum {
-    TM_PAGE_FLAG_WRITE_THROUGH = 1 << 0
-};
-
-/**
- * |struct tm_page| represents the transaction-local state of a block
- * of memory. The global counterpart is called |struct tm_frame|.
- */
-struct tm_page {
-    /** Block index and flag bits */
-    uintptr_t flags;
-
-    /** Lock state wrt. to frame lock */
-    struct picotm_rwstate rwstate;
-
-    /** Transaction-local buffer */
-    uint8_t buf[TM_BLOCK_SIZE];
-
-    /** Entry into allocator lists */
-    SLIST_ENTRY(tm_page) list;
-};
-
-static void
-tm_page_init(struct tm_page* page, struct tm_vmem* vmem, size_t block_index)
-{
-    page->flags = block_index << TM_BLOCK_SIZE_BITS;
-    picotm_rwstate_init(&page->rwstate);
-    memset(&page->list, 0, sizeof(page->list));
-}
-
-static void
-tm_page_uninit(struct tm_page* page)
-{
-    picotm_rwstate_uninit(&page->rwstate);
-}
-
-static size_t
-tm_page_block_index(const struct tm_page* page)
-{
-    return page->flags >> TM_BLOCK_SIZE_BITS;
-}
-
-static uintptr_t
-tm_page_address(const struct tm_page* page)
-{
-    return tm_page_block_index(page) * TM_BLOCK_SIZE;
-}
-
-static void*
-tm_page_buffer(struct tm_page* page)
-{
-    if (page->flags & TM_PAGE_FLAG_WRITE_THROUGH) {
-        return (void*)tm_page_address(page);
-    }
-    return page->buf; /* write-back is default mode */
-}
-
-static void
-tm_page_ld(struct tm_page* page, struct tm_vmem* vmem,
-           struct picotm_error* error)
-{
-    struct tm_frame* frame =
-        tm_vmem_acquire_frame_by_block(vmem, tm_page_block_index(page),
-                                       error);
-    if (picotm_error_is_set(error)) {
-        /* There's no legal way we should end up here! */
-        picotm_error_mark_as_non_recoverable(error);
-        return;
-    }
-
-    void* mem = tm_frame_buffer(frame);
-    memcpy(page->buf, mem, sizeof(page->buf));
-
-    tm_vmem_release_frame(vmem, frame);
-}
-
-static void
-tm_page_st(struct tm_page* page, struct tm_vmem* vmem,
-           struct picotm_error* error)
-{
-    struct tm_frame* frame =
-        tm_vmem_acquire_frame_by_block(vmem, tm_page_block_index(page),
-                                       error);
-    if (picotm_error_is_set(error)) {
-        /* There's no legal way we should end up here! */
-        picotm_error_mark_as_non_recoverable(error);
-        return;
-    }
-
-    void* buf = tm_frame_buffer(frame);
-    memcpy(buf, page->buf, sizeof(page->buf));
-
-    tm_vmem_release_frame(vmem, frame);
-}
-
-static void
-tm_page_xchg(struct tm_page* page, struct tm_vmem* vmem,
-             struct picotm_error* error)
-{
-    struct tm_frame* frame =
-        tm_vmem_acquire_frame_by_block(vmem, tm_page_block_index(page),
-                                       error);
-    if (picotm_error_is_set(error)) {
-        /* There's no legal way we should end up here! */
-        picotm_error_mark_as_non_recoverable(error);
-        return;
-    }
-
-    uint8_t buf[TM_BLOCK_SIZE];
-    void* fbuf = tm_frame_buffer(frame);
-    void* pbuf = page->buf;
-
-    memcpy( buf, fbuf, sizeof(buf));
-    memcpy(fbuf, pbuf, sizeof(buf));
-    memcpy(pbuf,  buf, sizeof(buf));
-
-    tm_vmem_release_frame(vmem, frame);
-}
-
-static bool
-tm_page_has_locked_frame(const struct tm_page* page)
-{
-    return picotm_rwstate_get_status(&page->rwstate) != PICOTM_RWSTATE_UNLOCKED;
-}
-
-static bool
-tm_page_has_rdlocked_frame(const struct tm_page* page)
-{
-    /* writer lock counts as implicit reader lock */
-    return picotm_rwstate_get_status(&page->rwstate) != PICOTM_RWSTATE_UNLOCKED;
-}
-
-static bool
-tm_page_has_wrlocked_frame(struct tm_page* page)
-{
-    return picotm_rwstate_get_status(&page->rwstate) == PICOTM_RWSTATE_WRLOCKED;
-}
-
-static void
-tm_page_try_rdlock_frame(struct tm_page* page, struct tm_vmem* vmem,
-                         struct picotm_error* error)
-{
-    struct tm_frame* frame =
-        tm_vmem_acquire_frame_by_block(vmem, tm_page_block_index(page),
-                                       error);
-    if (picotm_error_is_set(error)) {
-        return;
-    }
-    tm_frame_try_rdlock(frame, &page->rwstate, error);
-    if (picotm_error_is_set(error)) {
-        goto err_tm_frame_lock;
-    }
-    tm_vmem_release_frame(vmem, frame);
-
-    return;
-
-err_tm_frame_lock:
-    tm_vmem_release_frame(vmem, frame);
-}
-
-static void
-tm_page_try_wrlock_frame(struct tm_page* page, struct tm_vmem* vmem,
-                         struct picotm_error* error)
-{
-    struct tm_frame* frame =
-        tm_vmem_acquire_frame_by_block(vmem, tm_page_block_index(page),
-                                       error);
-    if (picotm_error_is_set(error)) {
-        return;
-    }
-    tm_frame_try_wrlock(frame, &page->rwstate, error);
-    if (picotm_error_is_set(error)) {
-        goto err_tm_frame_lock;
-    }
-    tm_vmem_release_frame(vmem, frame);
-
-    return;
-
-err_tm_frame_lock:
-    tm_vmem_release_frame(vmem, frame);
-}
-
-static void
-tm_page_unlock_frame(struct tm_page* page, struct tm_vmem* vmem,
-                     struct picotm_error* error)
-{
-    struct tm_frame* frame =
-        tm_vmem_acquire_frame_by_block(vmem, tm_page_block_index(page),
-                                       error);
-    if (picotm_error_is_set(error)) {
-        /* There's no legal way we should end up here! */
-        picotm_error_mark_as_non_recoverable(error);
-        return;
-    }
-
-    tm_frame_unlock(frame, &page->rwstate);
-    tm_vmem_release_frame(vmem, frame);
-}
-
-/*
- * |struct tm_vmem_tx|
- */
+#include "page.h"
 
 void
 tm_vmem_tx_init(struct tm_vmem_tx* vmem_tx, struct tm_vmem* vmem,
@@ -314,7 +103,7 @@ acquire_page_by_block(struct tm_vmem_tx* vmem_tx, size_t block_index,
         return NULL;
     }
 
-    tm_page_init(page, vmem_tx->vmem, block_index);
+    tm_page_init(page, block_index);
 
     /* The list of active pages is sorted by block index, so we
      * insert after the last page with a smaller block index, or
